# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Refresh user access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginPayloadDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Refresh user access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginPayloadDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/token": {
      "post": {
        "operationId": "AuthController_token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Refresh user access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginPayloadDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/checkEmail": {
      "post": {
        "operationId": "AuthController_checkEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Check if email is already in use",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Boolean"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/board": {
      "post": {
        "operationId": "BoardController_create",
        "parameters": [
          {
            "name": "project",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBoardDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create board",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "tags": [
          "Board"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "BoardController_findAll",
        "parameters": [
          {
            "name": "project",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Create board",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BoardDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "tags": [
          "Board"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/board/{id}": {
      "get": {
        "operationId": "BoardController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "project",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Get board",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "tags": [
          "Board"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "BoardController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "project",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBoardDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Get board",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "tags": [
          "Board"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "BoardController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "project",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Get board",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "tags": [
          "Board"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/epics": {
      "post": {
        "operationId": "EpicsController_create",
        "parameters": [
          {
            "name": "project",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEpicDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create board",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpicDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "tags": [
          "Epic"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "EpicsController_findAll",
        "parameters": [
          {
            "name": "project",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Create board",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EpicDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "tags": [
          "Epic"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/epics/{id}": {
      "get": {
        "operationId": "EpicsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "project",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Create board",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EpicDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "tags": [
          "Epic"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "EpicsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "project",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEpicDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create board",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EpicDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "tags": [
          "Epic"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "EpicsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "project",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Create board",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EpicDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          }
        },
        "tags": [
          "Epic"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Task Manager API",
    "description": "",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SignupDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "password"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "mobileNumber": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "userPermissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "projects": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "createdAt",
          "firstName",
          "lastName",
          "email",
          "mobileNumber",
          "isActive",
          "userPermissions",
          "roles",
          "projects"
        ]
      },
      "TokenPayloadDto": {
        "type": "object",
        "properties": {
          "expiresIn": {
            "type": "number"
          },
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "expiresIn",
          "accessToken",
          "refreshToken"
        ]
      },
      "LoginPayloadDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserDto"
          },
          "token": {
            "$ref": "#/components/schemas/TokenPayloadDto"
          }
        },
        "required": [
          "user",
          "token"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "CheckEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "Boolean": {
        "type": "object",
        "properties": {}
      },
...
      "ProjectDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "abbreviation": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "color": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "abbreviation",
          "description",
          "color"
        ]
      },
...
      "CreateBoardColumnDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "boardId": {
            "type": "number"
          },
          "taskStatus": {
            "type": "string",
            "enum": [
              "ToDo",
              "InProgress",
              "Done"
            ],
            "default": "ToDo"
          }
        },
        "required": [
          "name",
          "description",
          "position",
          "boardId",
          "taskStatus"
        ]
      },
      "CreateBoardDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateBoardColumnDto"
            }
          }
        },
        "required": [
          "name",
          "description",
          "position",
          "columns"
        ]
      },
      "EpicDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "projectId": {
            "type": "number"
          },
          "position": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "projectId",
          "position",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "TaskDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "issueTypeId": {
            "type": "number"
          },
          "issueType": {
            "$ref": "#/components/schemas/IssueTypeDto"
          },
          "epicId": {
            "type": "number"
          },
          "epic": {
            "$ref": "#/components/schemas/EpicDto"
          },
          "projectId": {
            "type": "number"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectDto"
          },
          "boardId": {
            "type": "number"
          },
          "board": {
            "$ref": "#/components/schemas/BoardDto"
          },
          "boardColumnId": {
            "type": "number"
          },
          "boardColumn": {
            "$ref": "#/components/schemas/BoardColumnDto"
          },
          "isBacklog": {
            "type": "boolean"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ],
            "default": "LOW"
          },
          "taskStatus": {
            "type": "string",
            "enum": [
              "ToDo",
              "InProgress",
              "Done"
            ],
            "default": "ToDo"
          },
          "assigneeId": {
            "type": "number"
          },
          "assignee": {
            "$ref": "#/components/schemas/UserDto"
          },
          "reporterId": {
            "type": "number"
          },
          "reporter": {
            "$ref": "#/components/schemas/UserDto"
          },
          "createdById": {
            "type": "number"
          },
          "createdBy": {
            "$ref": "#/components/schemas/UserDto"
          },
          "deletedById": {
            "type": "number"
          },
          "deletedBy": {
            "$ref": "#/components/schemas/UserDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "issueTypeId",
          "issueType",
          "epicId",
          "epic",
          "projectId",
          "project",
          "boardId",
          "board",
          "boardColumnId",
          "boardColumn",
          "isBacklog",
          "priority",
          "taskStatus",
          "assigneeId",
          "assignee",
          "reporterId",
          "reporter",
          "createdById",
          "createdBy",
          "deletedById",
          "deletedBy",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "BoardDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "projectId": {
            "type": "number"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectDto"
          },
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BoardColumnDto"
            }
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "position",
          "projectId",
          "project",
          "columns",
          "tasks",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "BoardColumnDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "boardId": {
            "type": "number"
          },
          "board": {
            "$ref": "#/components/schemas/BoardDto"
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskDto"
            }
          },
          "taskStatus": {
            "type": "string",
            "enum": [
              "ToDo",
              "InProgress",
              "Done"
            ],
            "default": "ToDo"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "position",
          "boardId",
          "board",
          "tasks",
          "taskStatus",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "UpdateBoardColumnDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "boardId": {
            "type": "number"
          },
          "taskStatus": {
            "type": "string",
            "enum": [
              "ToDo",
              "InProgress",
              "Done"
            ],
            "default": "ToDo"
          },
          "id": {
            "type": "number"
          }
        },
        "required": [
          "id"
        ]
      },
      "UpdateBoardDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "columns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateBoardColumnDto"
            }
          }
        }
      },
...
      "CreateEpicDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "position": {
            "type": "number"
          }
        },
        "required": [
          "name",
          "description",
          "position"
        ]
      },
      "UpdateEpicDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "id": {
            "type": "number"
          }
        },
        "required": [
          "id"
        ]
      }
    }
  }
}